IN-327 ref: Refactor folder structure
