--- src/app/[locale]/profile/[id]/followers/page.tsx
@@ -1,3 +1,0 @@
-export default function HomePage() {
-  return <div className='m-50 animate-bounce'>followers</div>
-}
