--- src/app/[locale]/profile/[id]/payments/page.tsx
@@ -1,11 +1,0 @@
-import { Payments } from '@/src/features/payments'
-
-export default async function PaymentsPage({
-  params,
-}: {
-  params: Promise<{ id: string }>
-}) {
-  const { id: userId } = await params
-
-  return <Payments userId={userId} />
-}
