--- src/features/payments/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './Payments'
