+++ src/features/profile/api/request.ts
@@ -10,12 +10,43 @@ export const GET_FOLLOWERS = gql(`
                 lastName
                 createdAt
                 userId
-               
             }
-                totalCount, 
-                pageSize, 
-                page, 
-                pagesCount
+            totalCount, 
+            pageSize, 
+            page, 
+            pagesCount
+        }
+    }
+`)
+
+export const GET_PAYMENTS = gql(`
+    query getPaymentsByUser($pageNumber: Int!, $pageSize: Int!, $Id: Int!,) {
+        getPaymentsByUser(pageNumber: $pageNumber, pageSize: $pageSize, userId: $Id,) {
+            pagesCount
+            page
+            pageSize
+            totalCount
+            items {
+                dateOfPayment
+                endDate
+                price 
+                paymentType
+                type
+            }
+        }
+    }
+`)
+
+export const GET_USER_FOTOS = gql(`
+    query getUserFotos($Id:Int!, $endCursorId: Int!) {
+        getPostsByUser(userId:$Id, endCursorId:$endCursorId){
+            pagesCount
+            pageSize
+            totalCount
+            items {
+                id
+                url
+            }
         }
     }
 `)
