+++ src/features/profile/model/index.ts
@@ -1,1 +1,4 @@
 export * from './types'
+export * from './usePaymentsList'
+export * from './useGetPaymentsColumns'
+export * from './useGetFotos'
