+++ src/features/profile/model/types.ts
@@ -7,3 +7,10 @@ export type Follower = {
   createdAt: string
   userId: number
 }
+
+import { GetPaymentsByUserQuery } from '@/src/shared/apolloClient/__generated__/graphql'
+
+export type PaymentItem =
+  GetPaymentsByUserQuery['getPaymentsByUser']['items'][number]
+
+export type PaymentItems = GetPaymentsByUserQuery['getPaymentsByUser']['items']
