+++ src/features/profile/model/usePaymentsList.ts
@@ -2,11 +2,11 @@ import { useEffect, useState } from 'react'
 import { useLazyQuery } from '@apollo/client'
 
 import { responseErrorHandler, usePagination } from '@/src/shared'
-import { GET_PAYMENTS } from '../api/request'
-import { PaymentItems } from './type'
+import { PaymentItems } from './types'
+import { GET_PAYMENTS } from '../api'
 
 export const usePaymentsList = (userId: string) => {
-  const [payments, setPayments] = useState<PaymentItems>()
+  const [payments, setPayments] = useState<PaymentItems>([])
   const [totalCount, setTotalCount] = useState<number>(0)
 
   const { currentPage, pageSize, onPageSize, onPageChange } = usePagination()
