+++ src/features/profile/ui/followers/index.ts
@@ -1,0 +1,1 @@
+export * from './Followers'
