+++ src/features/profile/ui/index.ts
@@ -1,1 +1,3 @@
 export * from './getTabs'
+export * from './payments'
+export * from './followers'
