+++ src/features/profile/ui/payments/index.ts
@@ -1,0 +1,1 @@
+export * from './PaymentsTable'
