+++ src/features/profile/ui/uploadedFotos/index.ts
